Multiple changes: Improved pickup function, improved calibration, almost working
Pickup function has been fixed almost completely. Droppnig off has also
been fixed. Robot is almost ready to play!

The calibration funciton is almost 100% perfect, it never misrecognizes
the markers on the side of the board.

A new function has been implimented that finds the ground height in
termes of the robots 'height' movement, which has been used in the
pickup and dropoff functions

Next? Time to get it playing checkers all the way! Move handling will be
done by waving your hand in front of the camera, promting the robot to
make a move.